refactor(DonutChart): update imports to use centralized config for geometry constants
